api/requirements.txt
anyio==3.5.0
asgiref==3.5.0
black==21.12b0
click==8.0.3
dataclasses-json==0.5.6
fastapi==0.73.0
h11==0.13.0
idna==3.3
marshmallow==3.14.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
psycopg2==2.9.3
pydantic==1.9.0
python-dotenv==0.19.2
sniffio==1.2.0
starlette==0.17.1
tomli==1.2.3
typing-inspect==0.7.1
typing_extensions==4.0.1
uvicorn==0.17.0
